Make Kiwix JS an Electron app
This commit brings in the www/ directory of Kiwix JS into this project.
The version of Kiwix JS that is included is
0b6d9b42ac73c1d84f790b252abd4c244c052fa6. Please note that attribution
and copyright information has NOT been updated. This is a simple
copy/paste of the www/ directory into this project. Kiwix JS can be
found here https://github.com/kiwix/kiwix-js. This information will be
updated accordingly before this repo is made 'public'.

As it stands, Kiwix JS is now a standalone Electron powered application.
To start it, run 'npm start' and it should run. Please note this has
only been 'tested' on macOS. Service worker support is very limited, so
please use the jQuery option.