Include electron-forge and switch to yarn
Electron-forge has been added to the project. This means that the
application can now be built and packaged with ease. Please note that I
have only tested this so far on macOS.

Electron-compile comes with electron-forge, this means that we get babel
and other such tools without having to do much (if any) configuration.
This is important as I intend to use as many modern language features as
possible. Most notably, my intention is to remove the need for requirejs
from this project and use ES6 module support.

I have included a yarn.lock and removed package-lock.json from the
project. I intend to use yarn, over npm, going forward. (This might be
subject to change...)