Include 'safe-node-app' dependency
This commit includes the SAFE Node API for the very first time. I have
also included the required changes to the package.json and main.js to
support connectivity with the SAFE Browser. In it's current state, it is
now possible to easily authenticate and communicate with the SAFE
Network.