Successfully authenticate app with SAFE Network
The app can now successfully connect, authenticate and communicate with
the SAFE Network. For now, the app will try and connect to the network
on launch. If successful, the app will check if it canAccessContainers()
then print some confirmation if successful.

I have toyed about with how best to do this for a few weeks. Eventually,
I settled on the approach used in the Web Hosting Manager example app.
For me personally, it is the most clean and easy to understand way of
doing this. I have pulled in some files from the WebHostingManager
(mostly boilerplate) and gave proper attribution in the files. The
'network.js' file contains the core logic that authenticates and
connects the app to the network. 'api.js' exists to give extra
functionality based on 'network.js'.

WebHostingManager:
github.com/maidsafe/safe_examples/tree/master/web_hosting_manager
(880a367bb6bdc0a61e33d0909e2089000148d0bd)