Add a very crude 'connect' button
Under 'configuration', there is now a simple button that requests auth
to the network. There is a spinner that will appear to show to the user
that network activity is happening. Once the user is authenticated with
the SAFE Network. A small label will change to inform the user that they
are now authenticated with the network.

To reiterate this is extremely crude. Work in progress.