Implement zim folder MD creation
This commit brings in the ability to create the zim folder MD that will
be emulated as NFS to store ZIM files. This MD exists in a users _public
container. The key is 'zim' and the value is a newly created public MD.

The MD housed at 'zim' is special. It has a zimFolderName which will be
publicly known and a type tag of '20010115'. This means that if another
user knows the zimFolderName, they will be able to read the data in this
MD. Effectively allowing them to read ZIM files that a given user has
uploaded.