Add an interface to create a users ZIM folder
This commit brings in a simple interface that can be used to create a
users ZIM folder. There is no validation on the form as of yet. Once a
user has created a folder, a small label will display as such. A label
will also display if the user doesn't have a ZIM folder.