Implement zim file uploading
This commit brings in the ability to upload zim files to the SAFE
Network. When given the local path to the file, and the file name,
uploader.js can upload the given file to the users 'zim' 'folder' in
their _public container. To do this, the 'zim' MD is accessed through
NFS emulation and the target file is placed inside the MD with the
specified fileName.

So far, the interface for this process isn't very good. Status of the
process is output to console for now and any errors that occur can be
seen there too. A better mechanism for displaying to the user how things
are progressing will be added in due course. Possible tweaks to the
chunk-size used for uploading will probably be made too. A chunk size
of 1000000 bytes has been used as this is what Maidsafe have used in
their examples.

This is a huge step towards achieving this projects goal. I have adapted
the solution used in the web_hosting_manager example application, proper
attribution has been given in the license header.