Implement zim file reading
This commit brings in the, very hacked together, ability to read a zim
file from the network. It is a very 'hard-coded' implementation of the
feature and exists as a proof-of-concept.

The app expects a zim file called 'test' to be in the users 'zimFolder'.
So to get this to work, you have to upload a zim file to your folder
called 'test'. Then, in the configuration panel use the 'load local
file' option and select the file you have just uploaded. You will see in
the console all the SAFE Network activity as it loads the file. You have
to select the file as a 'local file' to start kiwix-js off, but rest
assured the content displayed to the user is coming from the SAFE
Network. Some information is still sourced from the local file, such as
file size etc. This is a limitation I will address very soon.

The next steps are to remove the need for the local version of the file,
in combination with tweaking the implementation to support both SAFE
Network access and local access at the same time. This will be selected
through the use of a button/toggle.