Remove need for local zim file
This commit 'flushes out' the zim file reading implementation. An
interface has been added which now allows the user to specify which MD
to find the zim file, and what its name is. (Please note, the MD
specification still defaults to a zim folder in the current users
_public container that is called 'test'... This will be addressed soon.)
Once the information has been specified and the user hits 'Load', the
ZIMFile is built and the user is then able to browse the file. I have
tried as best as I can to leave the existing implementation untouched.
SAFE Network code is now tightly integrated into the existing code that
Kiwix-JS already had.

The need for the local version of the file to exist has been eliminated.
The information needed, file size, is now fetched from the network when
the ZIMFile is created.

The functionality to load a local file, what Kiwix-JS originally
provided, has returned. The use of the SAFE Network vs the local file
system is now selected through a radio button.