Allow the selection of any 'ZIM folder'
This commit removes the 'hard-coded' 'zimFolderName' value of 'test'.
The name of the ZIM folder specified by the user for file reading used
to be ignored and defaulted to 'test'. This limitation has been removed
and the name specified by the user is now respected. This means that
multiple collections of ZIM files on the network can now be accessed. I
have tested this with two different users so far, and both users can
access ZIM files in both of the 'ZIM folders'.