Fix Linux (Ubuntu) support
As highlighted by a member on the forums, Linux wasn't working... I
made a forum post and asked the question on how to fix this.
https://forum.safedev.org/t/uri-scheme-registration-on-ubuntu-linux/1438

The issue was to do with URI registration on Ubuntu. The program was
functional but was unable to properly 'hear' the 'auth-response' from
the SAFE Browser. This meant that you couldn't connect to the SAFE
Network. The changes in constants.js fix this issue, big thanks to
@hunterlester for suggesting the fix.

I have also changed the use of app.quit() to app.exit() in main.js...
This is to fix a weird error message that was popping up when SAFE Wiki
heard the 'auth-response' message on Linux (Ubuntu).

After including these changes, the app now functions correctly packaged
and when ran through the terminal on both macOS and on Linux. I still
need to validate Windows...