Retain 'zim folder' MD
Instead of 'fetching'/'generating' the 'zim folder' MD for every file
access, the MD is retained inside zimFile. Holding onto the MD means
that the computation needed during a file read has been reduced. Before
this improvement there was a lot of unnecessary repetition.

I have reduced the 'chatter' on the console during file reads so it is
more readable.